+++ CHANGELOG.md
@@ -11,6 +11,8 @@ title: Changelog
 * i18n: simplify formatter system and add JSDoc
 * i18n: remove country flags (it makes no sense for a language)
 * `<cc-button>`: Add a circle form when in hide-text with an image
+* New component:
+  * `<cc-pricing-table>`
 
 ## 6.4.0 (2021-04-21)